Corner case: if some of the nodes in the taskpool did not have to communicate with others, it's possible that they send termination detection messages before that taskpool was registered, triggering an assert because the message cannot be delivered to any process.
Usual solution to delay messages in a waiting list until
the taskpools are registered and become ready to follow
the algorithm is implemented, with some caution to add
atomic operations only at taskpool readiness (once per
taskpool), and if receiving messages when the taskpool
was not ready / did not exist.

Thanks to Joseph John for finding the use case creating
the issue.